+++ 1.4/Defs/ApparelLayerDefs/AdditionalLayersShield.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Defs>
+  <ApparelLayerDef>
+    <defName>Shield</defName>
+    <label>Shield</label>
+    <drawOrder>600</drawOrder>
+  </ApparelLayerDef>
+</Defs>
