+++ 1.4/Defs/ThingCategoryDefs/ShieldThingCategoryDefs.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Defs>
+    <ThingCategoryDef>
+      <defName>Shield</defName>
+      <label>shield</label>
+      <parent>Apparel</parent>
+    </ThingCategoryDef>
+
+
+</Defs>
