+++ 1.4/Patches/Arrow Please/WeaponPatch_AP.xml
@@ -1,0 +1,82 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Arrow Please</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Torch"]/weaponTags</xpath>
+   <value>
+	<li>Shield_Sidearm</li>
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Slinger"]/weaponTags</xpath>
+   <value>
+	<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+ 
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Tomahawk"]/weaponTags</xpath>
+   <value>
+	<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Blowgun"]/weaponTags</xpath>
+   <value>
+	<li>Shield_Sidearm</li>
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Crossbow"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+ 
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Flame_Crossbow"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Arbalest"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+   
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Flame_Arbalest"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+   
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Ranged_Repeatingcrossbow"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>  
+   
+<!-- Don't allow certain melee weapons to be used with shields -->   
+
+
+   </operations>
+   </match>
+</Operation>
+   
+</Patch>
