+++ 1.4/Patches/Core/PawnKindPatch_Core.xml
