+++ 1.4/Patches/Rimfire/WeaponPatch_Rimfire_SpacerRanged.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+    <Operation Class="PatchOperationFindMod">
+        <mods>
+            <li>Rimfire 2.5</li>
+        </mods>
+        <match Class="PatchOperationSequence">
+            <success>Always</success>
+            <operations>
+
+                <!-- Charge pistol -->
+                <li Class="PatchOperationAdd">
+                    <xpath>/Defs/ThingDef[defName="RF_Gun_ChargePistol"]/weaponTags</xpath>
+                    <value>
+                        <li>Shield_Sidearm</li>
+                    </value>
+                </li>
+
+            </operations>
+        </match>
+    </Operation>
+
+</Patch>
