+++ 1.4/Patches/Rimsenal/WeaponPatch_RimsenalVanilla.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Rimsenal - Enhanced Vanilla Pack</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+		<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Gun_ChargeSidearm"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Gun_SmartPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Gun_ThrowingAxes"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Gun_ThrowingClubs"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Gun_PistolHeavyRevolver"]</xpath>
+   <value>
+    <weaponTags>
+		<li>Shield_Sidearm</li>
+	</weaponTags>	
+   </value>
+   </li>
+   
+<!-- Don't allow certain melee weapons to be used with shields -->   
+   
+   </operations>
+   </match>
+</Operation>
+   
+</Patch>
