+++ 1.4/Patches/Rimsenal/WeaponPatch_Rimsenal_Federation.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Rimsenal - Federation Faction Pack</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="AD_Repose"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="AD_EMB"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>   
+   
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="ADOriginal_Negotiator"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="AD_Negotiator"]</xpath>
+   <value>
+		<weaponTags>
+			<li>Shield_Sidearm</li>
+		</weaponTags>
+   </value>
+   </li>   
+   
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="AD_AuxPistol"]/weaponTags</xpath>
+	<value>
+		<li>Shield_Sidearm</li>
+  </value>
+   </li>
+
+   </operations>
+   </match>
+</Operation>
+   
+</Patch>
