+++ 1.4/Patches/Rimsenal/WeaponPatch_Rimsenal_Feral.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Rimsenal - Feral Faction Pack</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Gun_Gnasher"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>
+   </value>
+   </li>
+
+   
+<!-- Don't allow certain melee weapons to be used with shields -->   
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="MeleeWeapon_FeralGlaive"]/weaponTags</xpath>
+   <value>
+		<li>Shield_NoSidearm</li>
+   </value>
+   </li> 
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="Sunderer"]/weaponTags</xpath>
+   <value>
+		<li>Shield_NoSidearm</li>
+   </value>
+   </li> 
+   
+   </operations>
+   </match>
+</Operation>
+   
+</Patch>
