+++ 1.4/Patches/Vanilla Weapons Expanded/WeaponPatch_VWE.xml
@@ -1,0 +1,95 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Vanilla Weapons Expanded</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_HandCannon"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_CombatHandgun"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>	
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_SawedOffShotgun"]/weaponTags</xpath>
+   <value>
+		<li>LightShield_Sidearm</li>	
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_LightSMG"]/weaponTags</xpath>
+   <value>
+		<li>LightShield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_SMG"]/weaponTags</xpath>
+   <value>
+		<li>LightShield_Sidearm</li>	
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_Flintlock"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Throwing_Knives"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Throwing_Rocks"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>	
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_ChargePistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_ChargeSMG"]/weaponTags</xpath>
+   <value>
+		<li>LightShield_Sidearm</li>		
+   </value>
+   </li>
+  
+<!-- Don't allow certain melee weapons to be used with shields -->   
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_MeleeWeapon_Halberd"]/weaponTags</xpath>
+   <value>
+		<li>Shield_NoSidearm</li>	
+   </value>
+   </li> 
+   
+		</operations>
+   </match>
+</Operation>
+   
+</Patch>
