+++ 1.4/Patches/Vanilla Weapons Expanded/WeaponPatch_VWEC.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Vanilla Weapons Expanded - Coilguns</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_GaussMagnum"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+ 
+		</operations>
+   </match>
+</Operation>
+   
+</Patch>
