+++ 1.4/Patches/Vanilla Weapons Expanded/WeaponPatch_VWEF.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Vanilla Weapons Expanded - Frontier</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWEFT_Gun_HeavyRevolver"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWEFT_Gun_VolcanicPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>	
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWEFT_Gun_FrontierPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>	
+   </value>
+   </li>
+
+		</operations>
+   </match>
+</Operation>
+   
+</Patch>
