+++ 1.4/Patches/Vanilla Weapons Expanded/WeaponPatch_VWEL.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Vanilla Weapons Expanded - Laser</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWEL_Gun_LaserPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWEL_Gun_SalvagedLaserPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+<!-- Allow certain ranged weapons to be used with light shields -->
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWEL_Gun_LaserSMG"]/weaponTags</xpath>
+   <value>
+		<li>LightShield_Sidearm</li>	
+   </value>
+   </li>
+   
+		</operations>
+   </match>
+</Operation>
+   
+</Patch>
