+++ 1.4/Patches/Vanilla Weapons Expanded/WeaponPatch_VWEM.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Vanilla Weapons Expanded - Makeshift</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_MakeshiftPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+<!-- Allow certain ranged weapons to be used with light shields -->
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_MakeshiftSMG"]/weaponTags</xpath>
+   <value>
+		<li>LightShield_Sidearm</li>	
+   </value>
+   </li>
+   
+		</operations>
+   </match>
+</Operation>
+   
+</Patch>
