+++ 1.4/Patches/Vanilla Weapons Expanded/WeaponPatch_VWENL.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Patch>
+
+	<Operation Class="PatchOperationFindMod">
+		<mods>
+			<li>Vanilla Weapons Expanded - Non-Lethal</li>
+		</mods>
+		<match Class="PatchOperationSequence">
+			<operations>
+	
+<!-- Allow certain ranged weapons to be used with shields -->
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_Taser"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>		
+   </value>
+   </li>
+
+   <li Class="PatchOperationAdd">
+   <xpath>/Defs/ThingDef[defName="VWE_Gun_RubberBulletPistol"]/weaponTags</xpath>
+   <value>
+		<li>Shield_Sidearm</li>	
+   </value>
+   </li>
+
+   
+		</operations>
+   </match>
+</Operation>
+   
+</Patch>
