+++ About/About.xml
@@ -7,6 +7,7 @@
 		<li>1.1</li>
 		<li>1.2</li>
 		<li>1.3</li>		
+		<li>1.4</li>		
 	</supportedVersions>
 	<packageId>n7huntsman.combatshields</packageId>
 	<modDependencies>
