+++ LoadFolders.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <loadFolders>  
+  <v1.4>    
+    <li>1.4</li>	
+    <li>/</li>	
+  </v1.4>
   <v1.3>    
     <li>1.3</li>	
     <li>/</li>
