+++ Source/MedievalShields/ColorableShield.cs
@@ -7,34 +7,8 @@ using Random = System.Random;
 
 namespace CombatShields
 {
-    public class Apparel_Shield : Apparel
+    public class ColorableShield : Apparel_Shield
     {
-        public static readonly SoundDef SoundAbsorbDamage = SoundDef.Named("PersonalShieldAbsorbDamage");
-        public static readonly SoundDef SoundBreak = SoundDef.Named("PersonalShieldBroken");
-
-        public bool colorable;
-        public Vector3 impactAngleVect;
-        public Material shieldMat;
-
-        public bool ShouldDisplay
-        {
-            get
-            {
-                if (Wearer.Dead || Wearer.InBed() || Wearer.Downed || Wearer.IsPrisonerOfColony)
-                {
-                    return false;
-                }
-
-                if (Wearer.IsColonist && !Wearer.Drafted && !Wearer.InAggroMentalState)
-                {
-                    return false;
-                }
-
-                return true;
-            }
-        }
-
-
         public override bool CheckPreAbsorbDamage(DamageInfo dinfo)
         {
             if (dinfo.Instigator == null)
@@ -74,6 +48,7 @@ namespace CombatShields
             return false;
         }
 
+
         [DebuggerHidden]
         public override IEnumerable<Gizmo> GetWornGizmos()
         {
@@ -85,13 +60,12 @@ namespace CombatShields
 
         public override void DrawWornExtras()
         {
-            // base.DrawWornExtras();
+            base.DrawWornExtras();
 
             if (!ShouldDisplay)
             {
                 return;
             }
-            // base.DrawWornExtras();
 
             var num = 0f;
             var vector = Wearer.Drawer.DrawPos;
@@ -126,52 +100,10 @@ namespace CombatShields
                 shieldMat = MaterialPool.MatFrom(def.graphicData.texPath);
             }
 
+            shieldMat.color = Stuff.stuffProps.color;
             var matrix = default(Matrix4x4);
             matrix.SetTRS(vector, Quaternion.AngleAxis(num, Vector3.up), s);
             Graphics.DrawMesh(MeshPool.plane10, matrix, shieldMat, 0);
-        }
-
-        [StaticConstructorOnStartup]
-        internal class Gizmo_ShieldStatus : Gizmo
-        {
-            private static readonly Texture2D FullTex =
-                SolidColorMaterials.NewSolidColorTexture(new Color(0.2f, 0.2f, 0.24f));
-
-            private static readonly Texture2D EmptyTex = SolidColorMaterials.NewSolidColorTexture(Color.clear);
-            public Apparel_Shield shield;
-
-            public override float GetWidth(float maxWidth)
-            {
-                return 140f;
-            }
-
-            public override GizmoResult GizmoOnGUI(Vector2 topLeft, float maxWidth, GizmoRenderParms parms)
-            {
-                var rect = new Rect(topLeft.x, topLeft.y, GetWidth(140f), 75f);
-                Widgets.DrawWindowBackground(rect);
-
-                var rect2 = rect.ContractedBy(6f);
-                var rect3 = rect2;
-
-                rect3.height = rect.height / 2f;
-
-                Text.Font = GameFont.Tiny;
-                Widgets.Label(rect3, shield.LabelCap);
-
-                var rect4 = rect2;
-
-                rect4.yMin = rect.y + (rect.height / 2f);
-                float num = shield.MaxHitPoints / shield.HitPoints;
-
-                Widgets.FillableBar(rect4, num, FullTex, EmptyTex, false);
-                Text.Font = GameFont.Small;
-                Text.Anchor = TextAnchor.MiddleCenter;
-
-                Widgets.Label(rect4, shield.HitPoints + " / " + shield.MaxHitPoints);
-                Text.Anchor = TextAnchor.UpperLeft;
-
-                return new GizmoResult(0);
-            }
         }
     }
 }
